requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.0
pydantic-settings>=2.0
httpx>=0.27.0
Pillow>=10.0
requests>=2.31
